backend/requirements.txt
django==6.0.3
djangorestframework==3.16.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
google-genai==1.10.0